+++ src/main/kotlin/db/tables/BlockedUsers.kt
@@ -1,14 +1,14 @@
 package com.neelkamath.omniChatBackend.db.tables
 
+import com.neelkamath.omniChatBackend.db.CursorType
 import com.neelkamath.omniChatBackend.db.ForwardPagination
 import com.neelkamath.omniChatBackend.db.accountsNotifier
-import com.neelkamath.omniChatBackend.graphql.routing.AccountEdge
-import com.neelkamath.omniChatBackend.graphql.routing.AccountsConnection
-import com.neelkamath.omniChatBackend.graphql.routing.BlockedAccount
-import com.neelkamath.omniChatBackend.graphql.routing.UnblockedAccount
+import com.neelkamath.omniChatBackend.graphql.routing.*
+import com.neelkamath.omniChatBackend.toLinkedHashSet
 import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
+import org.jetbrains.exposed.sql.SqlExpressionBuilder.greater
 import org.jetbrains.exposed.sql.transactions.transaction
 
 /** The users each user has blocked. */
@@ -16,6 +16,8 @@ object BlockedUsers : IntIdTable() {
     override val tableName = "blocked_users"
     private val blockerUserId: Column<Int> = integer("blocker_user_id").references(Users.id)
     private val blockedUserId: Column<Int> = integer("blocked_user_id").references(Users.id)
+
+    data class BlockedUserCursor(val blockedUserId: Int, val cursor: Cursor)
 
     /**
      * Has the [blockerUserId] block the [blockedUserId] if they haven't already. The [blockerUserId] will be notified
@@ -32,12 +34,47 @@ object BlockedUsers : IntIdTable() {
         accountsNotifier.publish(blockerUserId to BlockedAccount.build(blockedUserId))
     }
 
-    /** Reads the list of users the [userId] has blocked. */
-    fun read(userId: Int, pagination: ForwardPagination? = null): AccountsConnection {
-        val edges = transaction {
-            select(blockerUserId eq userId).map { AccountEdge.build(it[blockedUserId], it[BlockedUsers.id].value) }
+    /**
+     * Returns the users blocked by the [blockerUserId] as per the [pagination].
+     *
+     * The [BlockedUserCursor]s are sorted in ascending order of their [BlockedUserCursor.cursor].
+     */
+    fun readBlockedUsers(blockerUserId: Int, pagination: ForwardPagination? = null): LinkedHashSet<BlockedUserCursor> {
+        var op = BlockedUsers.blockerUserId eq blockerUserId
+        if (pagination?.after != null) op = op and (BlockedUsers.id greater pagination.after)
+        return transaction {
+            select(op)
+                .orderBy(BlockedUsers.id)
+                .let { if (pagination?.first == null) it else it.limit(pagination.first) }
+                .map { BlockedUserCursor(it[blockedUserId], cursor = it[BlockedUsers.id].value) }
+                .toLinkedHashSet()
         }
-        return AccountsConnection.build(edges.toSet(), pagination)
+    }
+
+    /** The [lastEdgeCursor] is the last [BlockedUsers.id] read by the [pagination] for the [blockerUserId]. */
+    fun readPageInfo(blockerUserId: Int, lastEdgeCursor: Int?, pagination: ForwardPagination? = null): PageInfo {
+        val startCursor = readCursor(blockerUserId, CursorType.START)
+        val endCursor = readCursor(blockerUserId, CursorType.END)
+        return PageInfo.build(lastEdgeCursor, startCursor, endCursor, pagination)
+    }
+
+    /**
+     * Reads the [type] of cursor for the [blockerUserId]. Returns `null` if the [blockerUserId] hasn't blocked any
+     * users.
+     */
+    private fun readCursor(blockerUserId: Int, type: CursorType): Int? {
+        val order = when (type) {
+            CursorType.END -> SortOrder.DESC
+            CursorType.START -> SortOrder.ASC
+        }
+        return transaction {
+            select(BlockedUsers.blockerUserId eq blockerUserId)
+                .orderBy(BlockedUsers.id, order)
+                .limit(1)
+                .firstOrNull()
+                ?.get(BlockedUsers.id)
+                ?.value
+        }
     }
 
     /**